+++ packages/skeleton-common/src/index.ts
@@ -2,5 +2,7 @@ export * from './themes';
 export * from './classes/accordion';
 export * from './classes/avatar';
 export * from './classes/combobox';
+export * from './classes/progress-linear';
 export * from './classes/rating-group';
+export * from './classes/switch';
 export * from './classes/tabs';