+++ packages/skeleton-svelte/src/index.ts
@@ -1,7 +1,9 @@
 export * from './components/accordion/index';
 export * from './components/avatar/index';
 export * from './components/combobox/index';
+export * from './components/progress-linear/index';
 export * from './components/rating-group/index';
+export * from './components/switch/index';
 export * from './components/tabs/index';
 
 export * from './hooks/use-list-collection';