+++ packages/skeleton-svelte/test/components/switch/switch.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+	import { Switch } from '@/index';
+</script>
+
+<Switch data-testid="root">
+	<Switch.HiddenInput data-testid="hidden-input" />
+	<Switch.Control data-testid="control">
+		<Switch.Thumb data-testid="thumb" />
+	</Switch.Control>
+	<Switch.Label data-testid="label" />
+</Switch>